# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Note Taking App API",
    "version": "1.0.0",
    "description": "A comprehensive note-taking application with user authentication and note management features",
    "contact": {
      "name": "API Support",
      "email": "support@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    },
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token for authentication. Format: Bearer <token>"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "next-auth.session-token",
        "description": "NextAuth session cookie authentication"
      }
    },
    "schemas": {
...
      "BlockDataSchema": {
        "oneOf": [
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          },
          {
            "type": "object"
          }
        ]
      },
      "EditorBlockSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique block identifier"
          },
          "type": {
            "type": "string",
            "enum": [
              "header",
              "paragraph",
              "list",
              "quote",
              "delimiter"
            ],
            "description": "Block type"
          },
          "data": {
            "allOf": [
              {
                "$ref": "#/components/schemas/BlockDataSchema"
              }
            ],
            "description": "Block-specific data"
          }
        },
        "required": [
          "id",
          "type",
          "data"
        ]
      },
      "EditorContentSchema": {
        "type": "object",
        "properties": {
          "time": {
            "type": "number",
            "description": "Content creation/modification timestamp"
          },
          "blocks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EditorBlockSchema"
            },
            "description": "Array of content blocks"
          },
          "version": {
            "type": "string",
            "description": "Editor.js version"
          }
        },
        "required": [
          "time",
          "blocks",
          "version"
        ]
      },
...
    },
    "responses": {
      "400": {
        "description": "Bad Request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "Invalid request parameters"
                },
                "code": {
                  "type": "string",
                  "example": "400"
                }
              }
            }
          }
        }
      },
...
      "500": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "example": "An unexpected error occurred"
                },
                "code": {
                  "type": "string",
                  "example": "500"
                }
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/notes": {
      "get": {
        "operationId": "get-notes",
        "summary": "Get notes with optional filters",
        "description": "Retrieve notes with optional filters for title, content, tags, and archive status",
        "tags": [
          "Notes"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "schema": {
              "type": "string",
              "description": "Search query"
            },
            "required": false,
            "description": "Search query"
          },
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "array",
              "description": "Filter by tags"
            },
            "required": false,
            "description": "Filter by tags"
          },
          {
            "in": "query",
            "name": "isArchived",
            "schema": {
              "type": "boolean",
              "description": "Filter by archive status"
            },
            "required": false,
            "description": "Filter by archive status"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "description": "Page number"
            },
            "required": false,
            "description": "Page number"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "description": "Items per page"
            },
            "required": false,
            "description": "Items per page"
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Notes retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Response message"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "message",
                    "data",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      },
      "post": {
        "operationId": "post-notes",
        "summary": "Create note",
        "description": "Create a new note with title, content, and optional tags",
        "tags": [
          "Notes"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "description": "Note title"
                  },
                  "content": {
                    "allOf": [
                      {
                        "$ref": "#/components/schemas/EditorContentSchema"
                      }
                    ],
                    "description": "Note content in Editor.js format"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "nullable": true,
                    "description": "Array of tag names"
                  }
                },
                "required": [
                  "title",
                  "content"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Note created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Response message"
                    },
                    "note": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "message",
                    "note",
                    "note"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
    "/tags": {
      "get": {
        "operationId": "get-tags",
        "summary": "Get tags",
        "description": "Retrieve all user tags with note counts and usage statistics",
        "tags": [
          "Tags"
        ],
        "parameters": [],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tags retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Response message"
                    },
                    "data": {
                      "type": "object"
                    }
                  },
                  "required": [
                    "message",
                    "data",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400"
          },
          "500": {
            "$ref": "#/components/responses/500"
          }
        }
      }
    },
...
  }
}